requirements.txt
attrs==19.3.0
biopython==1.76
clipkit==0.0.1
more-itertools==8.2.0
numpy==1.18.2
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
six==1.14.0
wcwidth==0.1.9
